[refactor]: split patient timeline service into access, build and narrative helpers
